requirements.txt
asgiref==3.2.10
autopep8==1.5.4
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
omnitools==0.0.94
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
toml==0.10.2